Issue 67.4: extract forward-request builders from afxdp.rs
Final step of Issue 67. afxdp.rs goes from 862 → 753 prod LOC by
relocating three forward-request builders into a new sibling
module afxdp/forward_request.rs:

- should_install_local_reverse_session — predicate that decides
  whether to pre-install reverse-direction session entry locally
- build_live_forward_request — packs ForwardingResolution +
  SessionMetadata into LiveForwardRequest descriptor
- build_live_forward_request_from_frame — frame-input wrapper

All three widened from file-private `fn` to `pub(super) fn` so
afxdp.rs can call them via `use forward_request::{...};`.

Also: cleans up imports in afxdp.rs that were previously firing
`unused_imports` warnings — `learn_dynamic_neighbor` (#1071's bundle)
and now `update_last_resolution` and `build_live_forward_request`
are all only consumed by tests (afxdp/forwarding.rs cfg-test
sections, afxdp/tests.rs, afxdp/frame/tests.rs). Gated their
imports behind `#[cfg(test)]` so non-test builds don't trip the
warnings. Warning count is back to the 92-warning baseline.

afxdp.rs adds:
    mod forward_request;
    use forward_request::{
        build_live_forward_request_from_frame,
        should_install_local_reverse_session,
    };
    #[cfg(test)]
    use forward_request::build_live_forward_request;

Pure relocation. Bodies byte-for-byte identical.

cargo test --release: 865 passed, 0 failed.

Cumulative Issue 67 result:
- afxdp.rs: 1,454 → 753 prod LOC
- afxdp/session_delta.rs:        197 (Issue 67.1, #1068)
- afxdp/neighbor_dispatch.rs:    218 (Issue 67.2, #1070)
- afxdp/disposition.rs:          254 (Issue 67.3, #1071)
- afxdp/forward_request.rs:      134 (Issue 67.4, this PR)
